// backend/src/entity/enums.ts
export enum GroupType {
  HOUSEHOLD = "household",
  PERSONAL = "personal",
}

export enum MemberRole {
  ADMIN = "admin",
  MEMBER = "member",
}

export enum SplitType {
  EQUAL = "equal",
  EXACT = "exact",
  PERCENTAGE = "percentage",
  SHARES = "shares",
}

export enum RecurInterval {
  WEEKLY = "weekly",
  MONTHLY = "monthly",
  YEARLY = "yearly",
}